+++ src/core/vfs-fonts/index.ts
@@ -20,23 +20,4 @@
  *
  */
 
-import {AjfBaseWidgetComponent, AjfColumnWidgetInstance} from '@ajf/core/reports';
-import {
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component,
-  ElementRef,
-  ViewEncapsulation
-} from '@angular/core';
-
-@Component({
-  templateUrl: 'column-widget.html',
-  styleUrls: ['column-widget.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None,
-})
-export class AjfColumnWidgetComponent extends AjfBaseWidgetComponent<AjfColumnWidgetInstance> {
-  constructor(cdr: ChangeDetectorRef, el: ElementRef) {
-    super(cdr, el);
-  }
-}
+export * from './public-api';