+++ src/core/vfs-fonts/vfs-fonts.ts
@@ -19,3 +19,7 @@
  * If not, see http://www.gnu.org/licenses/.
  *
  */
+
+import {vfsFonts as vfsFontsLib} from './vfs-fonts-lib';
+
+export const vfsFonts = {...vfsFontsLib};