+++ src/ionic/reports/public-api.ts
@@ -21,11 +21,9 @@
  */
 
 export * from './chart-widget';
-export * from './column-widget';
 export * from './formula-widget';
 export * from './image-container-widget';
 export * from './image-widget';
-export * from './layout-widget';
 export * from './map-widget';
 export * from './page-break-widget';
 export * from './report';
@@ -33,4 +31,3 @@ export * from './reports-module';
 export * from './table-widget';
 export * from './text-widget';
 export * from './widget';
-export * from './widget-service';